requirements.txt
streamlit
pdf2image
pytesseract
opencv-python
numpy
pandas
pillow
joblib
plotly